Renamed dense_vec_serialization.cpp
